Title for registered users should be blank if the user does not have an admin role
