+++ src/i18n/en/translation.en.json
@@ -30,67 +30,83 @@
       "authorization-credentials": {
         "GlobalAdmin": {
           "name": "Global Admin",
-          "order": 0
+          "order": 0,
+          "hidden": false
         },
         "GlobalAdminCommunity": {
           "name": "Global Community Admin",
-          "order": 1
+          "order": 1,
+          "hidden": false
         },
         "GlobalAdminChallenges": {
           "name": "Global Challenges Admin",
-          "order": 2
+          "order": 2,
+          "hidden": false
         },
         "GlobalRegistered": {
           "name": "Registered",
-          "order": 999
+          "order": 999,
+          "hidden": true
         },
         "EcoverseAdmin": {
           "name": "Ecoverse Admin",
-          "order": 3
+          "order": 3,
+          "hidden": false
         },
         "EcoverseHost": {
           "name": "Ecoverse Host",
-          "order": 4
+          "order": 4,
+          "hidden": true
         },
         "EcoverseMember": {
           "name": "Ecoverse Member",
-          "order": 5
+          "order": 5,
+          "hidden": true
         },
         "ChallengeAdmin": {
           "name": "Challenge Admin",
-          "order": 6
+          "order": 6,
+          "hidden": false
         },
         "ChallengeLead": {
           "name": "Challenge Lead",
-          "order": 7
+          "order": 7,
+          "hidden": true
         },
         "ChallengeMember": {
           "name": "Challenge Member",
-          "order": 8
+          "order": 8,
+          "hidden": true
         },
         "OpportunityMember": {
           "name": "Opportunity Member",
-          "order": 9
+          "order": 9,
+          "hidden": true
         },
         "OrganisationAdmin": {
           "name": "Organisation Admin",
-          "order": 10
+          "order": 10,
+          "hidden": false
         },
         "OrganisationMember": {
           "name": "Organisation Member",
-          "order": 11
+          "order": 11,
+          "hidden": true
         },
         "UserSelfManagement": {
           "name": "User Self Management",
-          "order": 12
+          "order": 12,
+          "hidden": true
         },
         "CommunityMember": {
           "name": "Community Member",
-          "order": 13
+          "order": 13,
+          "hidden": true
         },
         "UserGroupMember": {
           "name": "User Group Member",
-          "order": 14
+          "order": 14,
+          "hidden": true
         }
       }
     }
